+++ classes/hypeJunction/Notifications/AddHtmlEmailPart.php
@@ -20,7 +20,7 @@ class AddHtmlEmailPart {
 		$message = $hook->getValue();
 		/* @var $message Message */
 
-		if (elgg_get_plugin_setting('enable_html_emails', 'hypeNotifications')) {
+		if (elgg_get_plugin_setting('enable_html_emails', 'hypenotifications')) {
 
 			$html_body = elgg_view('notifications/wrapper/html', [
 				'email' => $hook->getParam('email'),